+++ tel-testdata/tel-testdata-integrations/src/test/groovy/dk/sunepoulsen/tel/testdata/module/integrator/TelTestDataIntegratorSpec.groovy
@@ -4,6 +4,7 @@ import dk.sunepoulsen.tel.testdata.module.integrator.model.DataPointDataSet
 import dk.sunepoulsen.tel.testdata.module.integrator.model.DataPointDataSetConstraints
 import dk.sunepoulsen.tes.rest.integrations.TechEasySolutionsClient
 import dk.sunepoulsen.tes.rest.integrations.exceptions.ClientBadRequestException
+import dk.sunepoulsen.tes.rest.integrations.exceptions.ClientNotFoundException
 import dk.sunepoulsen.tes.rest.models.RangeModel
 import dk.sunepoulsen.tes.rest.models.ServiceErrorModel
 import spock.lang.Specification
@@ -28,8 +29,8 @@ class TelTestDataIntegratorSpec extends Specification {
                 name: 'name',
                 description: 'description',
                 constraints: new DataPointDataSetConstraints(
-                    xValues: new RangeModel<Double>(min: 1.0, max: 200.0),
-                    yValues: new RangeModel<Double>(min: 1.0, max: 200.0),
+                    xValues: new RangeModel<BigDecimal>(min: 1.0, max: 200.0),
+                    yValues: new RangeModel<BigDecimal>(min: 1.0, max: 200.0),
                     quantity: new RangeModel<Integer>(min: 50, max: 20000)
                 )
             )
@@ -39,8 +40,8 @@ class TelTestDataIntegratorSpec extends Specification {
                 name: 'name',
                 description: 'description',
                 constraints: new DataPointDataSetConstraints(
-                    xValues: new RangeModel<Double>(min: 1.0, max: 200.0),
-                    yValues: new RangeModel<Double>(min: 1.0, max: 200.0),
+                    xValues: new RangeModel<BigDecimal>(min: 1.0, max: 200.0),
+                    yValues: new RangeModel<BigDecimal>(min: 1.0, max: 200.0),
                     quantity: new RangeModel<Integer>(min: 50, max: 20000)
                 )
             )
@@ -61,8 +62,8 @@ class TelTestDataIntegratorSpec extends Specification {
                 name: 'name',
                 description: 'description',
                 constraints: new DataPointDataSetConstraints(
-                    xValues: new RangeModel<Double>(min: 1.0, max: 200.0),
-                    yValues: new RangeModel<Double>(min: 1.0, max: 200.0),
+                    xValues: new RangeModel<BigDecimal>(min: 1.0, max: 200.0),
+                    yValues: new RangeModel<BigDecimal>(min: 1.0, max: 200.0),
                     quantity: new RangeModel<Integer>(min: 50, max: 20000)
                 )
             )
@@ -79,6 +80,47 @@ class TelTestDataIntegratorSpec extends Specification {
                 )))
             }
             ClientBadRequestException ex = thrown(ClientBadRequestException)
+            ex.serviceError.code == 'code'
+            ex.serviceError.param == 'param'
+            ex.serviceError.message == 'message'
+    }
+
+    void "GET Data Point DataSet with OK"() {
+        given:
+            DataPointDataSet expected = new DataPointDataSet(
+                id: 1L,
+                name: 'name',
+                description: 'description',
+                constraints: new DataPointDataSetConstraints(
+                    xValues: new RangeModel<BigDecimal>(min: 1.0, max: 200.0),
+                    yValues: new RangeModel<BigDecimal>(min: 1.0, max: 200.0),
+                    quantity: new RangeModel<Integer>(min: 50, max: 20000)
+                )
+            )
+
+        when:
+            DataPointDataSet result = this.sut.get(1L).blockingGet()
+
+        then:
+            1 * httpClient.get('/datasets/data-points/1', DataPointDataSet) >> CompletableFuture.supplyAsync {
+                expected
+            }
+            result == expected
+    }
+
+    void "GET Data Point DataSet with error"() {
+        when:
+            this.sut.get(1L).blockingGet()
+
+        then:
+            1 * httpClient.get('/datasets/data-points/1', DataPointDataSet) >> CompletableFuture.supplyAsync {
+                throw new ExecutionException("message", new ClientNotFoundException(Mock(HttpResponse), new ServiceErrorModel(
+                    code: 'code',
+                    param: 'param',
+                    message: 'message'
+                )))
+            }
+            ClientNotFoundException ex = thrown(ClientNotFoundException)
             ex.serviceError.code == 'code'
             ex.serviceError.param == 'param'
             ex.serviceError.message == 'message'