+++ tel-testdata/tel-testdata-module/src/main/java/dk/sunepoulsen/tel/testdata/module/service/domains/datapoints/DataPointsController.java
@@ -2,10 +2,9 @@ package dk.sunepoulsen.tel.testdata.module.service.domains.datapoints;
 
 import dk.sunepoulsen.tel.testdata.module.integrator.model.DataPointDataSet;
 import dk.sunepoulsen.tes.springboot.rest.logic.exceptions.LogicException;
+import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequestMapping(DataPointsOperations.DATASETS_DATA_POINTS_ENDPOINT_PATH)
@@ -19,7 +18,7 @@ class DataPointsController implements DataPointsOperations {
     }
 
     @PostMapping
-    public DataPointDataSet createDataPointDataSet(DataPointDataSet dataPointDataSet) {
+    public DataPointDataSet create(DataPointDataSet dataPointDataSet) {
         try {
             DataPointDataSet result = this.logicService.create(dataPointDataSet);
             this.logicService.generateDataPoints(result);
@@ -28,4 +27,14 @@ class DataPointsController implements DataPointsOperations {
             throw ex.mapApiException();
         }
     }
+
+    @GetMapping("/{id}")
+    public DataPointDataSet get(@Valid @PathVariable("id") Long id) {
+        try {
+            return logicService.get(id);
+        } catch (LogicException ex) {
+            throw ex.mapApiException();
+        }
+    }
+
 }