+++ tel-testdata/CHANGELOG.md
@@ -21,6 +21,8 @@ For release-dates, use date-format: YYYY-MM-DD
 - [TESENTLABS-27](https://sunepoulsen.atlassian.net/browse/TESENTLABS-27): Manage data sets in tel-testdata
   - Add api model for the endpoints `POST /datasets/data-points`
   - Add database migration for data sets with data points.
+  - Implemented `POST /datasets/data-points`
+  - Implemented `GET /datasets/data-points/{id}`
 
 ### Fixed